Criação do controller home para página inicial, ajustes nos perfis (prod, dev e test)

+++ src/main/java/br/com/emersonmorgado/aluraflix/aluraflix/controller/CategoriaController.java
@@ -11,6 +11,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -78,6 +79,9 @@ public class CategoriaController {
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<String> remover(@PathVariable Long id){
+		if ( id < 3) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Sem permissão para remover este registro");
+		}
 		Optional<Categoria> categoria = categoriaService.findById(Long.valueOf(id));
 		if(categoria.isPresent()) {
 			categoriaService.remove(categoria.get());
@@ -87,7 +91,10 @@ public class CategoriaController {
 	}
 	@PutMapping("/{id}")
 	@Transactional
-	public ResponseEntity<CategoriaDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizaCategoriaForm form){
+	public ResponseEntity<Object> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizaCategoriaForm form){
+		if ( id < 5) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Sem permissão para atualizar este registro");
+		}
 		Optional<Categoria> categoria = categoriaService.findById(id);
 		if(categoria.isPresent()) {
 			CategoriaDto categoriaDto = categoriaService.atualizar(categoria.get(), form);
